part-10/python/requirements.txt
selenium
pynput
google-api-python-client
google-cloud-storage
Pillow